adding the compute_score and get_string function
In main(), the program prompts the two players for their words using the get_string() function. These values are stored inside variables named word1 and word2.

In compute_score(), the program will compute, using the POINTS array, and return the score for the string argument. Characters that are not letters will be given zero points, and uppercase and lowercase letters will be given the same point values.